Refactor conditions.go into its own package
conditions.go was a giant file with a lot going on. This PR splits it up into its own
packagage.

Additionally, conditions.go was checking the configured conditions with conditional branching both at
config load time and once per event. This patch removes the redundant per-event checks.

There may be a small perf boost with this patch as well.
