+++ libbeat/conditions/has_fields.go
@@ -15,13 +15,29 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package processors
+package conditions
 
-import (
-	"github.com/elastic/beats/libbeat/common"
-)
+import "fmt"
 
-type PluginConfig []map[string]*common.Config
+// HasFields is a Condition for checking field existence.
+type HasFields []string
 
-// fields that should be always exported
-var MandatoryExportedFields = []string{"type"}
+// NewHasFieldsCondition builds a new HasFields checking the given list of fields.
+func NewHasFieldsCondition(fields []string) (hasFieldsCondition HasFields) {
+	return HasFields(fields)
+}
+
+// Check determines whether the given event matches this condition
+func (c HasFields) Check(event ValuesMap) bool {
+	for _, field := range c {
+		_, err := event.GetValue(field)
+		if err != nil {
+			return false
+		}
+	}
+	return true
+}
+
+func (c HasFields) String() string {
+	return fmt.Sprintf("has_fields: %v", []string(c))
+}
