+++ metricbeat/autodiscover/appender/kubernetes/token/token.go
@@ -25,8 +25,8 @@ import (
 	"github.com/elastic/beats/libbeat/common"
 	"github.com/elastic/beats/libbeat/common/bus"
 	"github.com/elastic/beats/libbeat/common/cfgwarn"
+	"github.com/elastic/beats/libbeat/conditions"
 	"github.com/elastic/beats/libbeat/logp"
-	"github.com/elastic/beats/libbeat/processors"
 )
 
 func init() {
@@ -35,7 +35,7 @@ func init() {
 
 type tokenAppender struct {
 	TokenPath string
-	Condition *processors.Condition
+	Condition conditions.Condition
 }
 
 // NewTokenAppender creates a token appender that can append a bearer token required to authenticate with
@@ -49,10 +49,10 @@ func NewTokenAppender(cfg *common.Config) (autodiscover.Appender, error) {
 		return nil, fmt.Errorf("unable to unpack config due to error: %v", err)
 	}
 
-	var cond *processors.Condition
+	var cond conditions.Condition
 	if conf.ConditionConfig != nil {
 		// Attempt to create a condition. If fails then report error
-		cond, err = processors.NewCondition(conf.ConditionConfig)
+		cond, err = conditions.NewCondition(conf.ConditionConfig)
 		if err != nil {
 			return nil, fmt.Errorf("unable to create condition due to error: %v", err)
 		}
